feat(database/offline): implement QueryTableTimeRange for bigquery
