+++ internal/database/offline/bigquery/store.go
@@ -2,12 +2,9 @@ package bigquery
 
 import (
 	"context"
-	"fmt"
 
 	"cloud.google.com/go/bigquery"
 	"github.com/oom-ai/oomstore/pkg/errdefs"
-	"github.com/spf13/cast"
-	"google.golang.org/api/iterator"
 	"google.golang.org/api/option"
 
 	"github.com/oom-ai/oomstore/internal/database/dbutil"
@@ -42,54 +39,6 @@ func (db *DB) Ping(ctx context.Context) error {
 	q := db.Client.Query("SELECT 1")
 	_, err := q.Read(ctx)
 	return errdefs.WithStack(err)
-}
-
-func (db *DB) TableSchema(ctx context.Context, opt offline.TableSchemaOpt) (*types.DataTableSchema, error) {
-	q := fmt.Sprintf(`SELECT column_name, data_type FROM %s.INFORMATION_SCHEMA.COLUMNS WHERE table_name = "%s"`, db.datasetID, opt.TableName)
-	rows, err := db.Query(q).Read(ctx)
-	if err != nil {
-		return nil, errdefs.WithStack(err)
-	}
-	var schema types.DataTableSchema
-	for {
-		recordMap := make(map[string]bigquery.Value)
-		err = rows.Next(&recordMap)
-		if err == iterator.Done {
-			break
-		}
-		if err != nil {
-			return nil, errdefs.WithStack(err)
-		}
-		valueType, err := dbutil.ValueType(Backend, cast.ToString(recordMap["data_type"]))
-		if err != nil {
-			return nil, err
-		}
-		schema.Fields = append(schema.Fields, types.DataTableFieldSchema{
-			Name:      cast.ToString(recordMap["column_name"]),
-			ValueType: valueType,
-		})
-	}
-	if opt.CheckTimeRange {
-		q = fmt.Sprintf(`
-		SELECT
-			MIN(%s) AS %s,
-			MAX(%s) AS %s
-		FROM %s.%s`, "unix_milli", "min_unix_milli", "unix_milli", "max_unix_milli", db.datasetID, opt.TableName)
-		rows, err = db.Query(q).Read(ctx)
-		if err != nil {
-			return nil, errdefs.WithStack(err)
-		}
-		recordMap := make(map[string]bigquery.Value)
-		err = rows.Next(&recordMap)
-		if err != nil {
-			return nil, errdefs.WithStack(err)
-		}
-		schema.TimeRange = types.DataTableTimeRange{
-			MinUnixMilli: cast.ToInt64(recordMap["min_unix_milli"]),
-			MaxUnixMilli: cast.ToInt64(recordMap["max_unix_milli"]),
-		}
-	}
-	return &schema, nil
 }
 
 func (db *DB) Snapshot(ctx context.Context, opt offline.SnapshotOpt) error {
